requirements.txt
qq-botpy
langchain
langchain-openai
matplotlib
pillow